feat(ui): Improved history matching
The math was getting a bit out of hand in ui::UI, so I updated it
(multiply matching score by the history), and moved to a function inside
`Application` (corrected_score).

Now I use this corrected score in `Application.cmp`, which and in
`ui:UI` I only have to update `app.score`. (Bonus: the UI shows the
actual matching score, and not the corrected one)